refactor: 추가 parser 클래스
