+++ src/main/java/baseball/controller/GameController.java
@@ -1,6 +1,7 @@
 package baseball.controller;
 
 import baseball.domain.GameService;
+import baseball.utill.Parser;
 import baseball.utill.Validation;
 import camp.nextstep.edu.missionutils.Console;
 
@@ -10,6 +11,7 @@ import java.util.List;
 public class GameController {
 
     GameService service = new GameService();
+    Parser parser = new Parser();
     Validation validation = new Validation();
     List<Integer> computerNumbers = new ArrayList<>();
 
@@ -25,7 +27,7 @@ public class GameController {
             if (computerNumbers.isEmpty()) computerNumbers = service.generateNumbers();
             System.out.println("숫자를 입력해주세요.");
             String input = Console.readLine();
-            List<Integer> userNumbers = service.parseInputToList(input);
+            List<Integer> userNumbers = parser.parseInputToList(input);
             if (validation.isInvalid(userNumbers)) throw new IllegalArgumentException("올바르지 않은 입력값입니다.");
             int[] result = service.compare(computerNumbers, userNumbers);
             service.printResult(result);
